+++ src/dbxml.c
@@ -129,27 +129,27 @@ void xmlpercentile(const interfaceinfo *ifaceinfo)
 	printf("   <entries><seen>%" PRIu32 "</seen><expected>%" PRIu32 "</expected><missing>%" PRIu32 "</missing></entries>\n", pdata.count, pdata.countexpectation, pdata.countexpectation-pdata.count);
 
 	printf("   <minimum>");
-	printf("<rx_bytes_per_second>%" PRIu64 "</rx_bytes_per_second>", (uint64_t)(pdata.minrx / (double)300));
-	printf("<tx_bytes_per_second>%" PRIu64 "</tx_bytes_per_second>", (uint64_t)(pdata.mintx / (double)300));
-	printf("<total_bytes_per_second>%" PRIu64 "</total_bytes_per_second>", (uint64_t)(pdata.min / (double)300));
+	printf("<rx_bytes_per_second>%" PRIu64 "</rx_bytes_per_second>", (uint64_t)((double)pdata.minrx / (double)300));
+	printf("<tx_bytes_per_second>%" PRIu64 "</tx_bytes_per_second>", (uint64_t)((double)pdata.mintx / (double)300));
+	printf("<total_bytes_per_second>%" PRIu64 "</total_bytes_per_second>", (uint64_t)((double)pdata.min / (double)300));
 	printf("</minimum>\n");
 
 	printf("   <average>");
-	printf("<rx_bytes_per_second>%" PRIu64 "</rx_bytes_per_second>", (uint64_t)(pdata.sumrx / (double)(pdata.count * 300)));
-	printf("<tx_bytes_per_second>%" PRIu64 "</tx_bytes_per_second>", (uint64_t)(pdata.sumtx / (double)(pdata.count * 300)));
-	printf("<total_bytes_per_second>%" PRIu64 "</total_bytes_per_second>", (uint64_t)(pdata.sumrx + pdata.sumtx / (double)(pdata.count * 300)));
+	printf("<rx_bytes_per_second>%" PRIu64 "</rx_bytes_per_second>", (uint64_t)((double)pdata.sumrx / (double)(pdata.count * 300)));
+	printf("<tx_bytes_per_second>%" PRIu64 "</tx_bytes_per_second>", (uint64_t)((double)pdata.sumtx / (double)(pdata.count * 300)));
+	printf("<total_bytes_per_second>%" PRIu64 "</total_bytes_per_second>", (uint64_t)((double)pdata.sumrx + (double)pdata.sumtx / (double)(pdata.count * 300)));
 	printf("</average>\n");
 
 	printf("   <maximum>");
-	printf("<rx_bytes_per_second>%" PRIu64 "</rx_bytes_per_second>", (uint64_t)(pdata.maxrx / (double)300));
-	printf("<tx_bytes_per_second>%" PRIu64 "</tx_bytes_per_second>", (uint64_t)(pdata.maxtx / (double)300));
-	printf("<total_bytes_per_second>%" PRIu64 "</total_bytes_per_second>", (uint64_t)(pdata.max / (double)300));
+	printf("<rx_bytes_per_second>%" PRIu64 "</rx_bytes_per_second>", (uint64_t)((double)pdata.maxrx / (double)300));
+	printf("<tx_bytes_per_second>%" PRIu64 "</tx_bytes_per_second>", (uint64_t)((double)pdata.maxtx / (double)300));
+	printf("<total_bytes_per_second>%" PRIu64 "</total_bytes_per_second>", (uint64_t)((double)pdata.max / (double)300));
 	printf("</maximum>\n");
 
 	printf("   <95th_percentile>");
-	printf("<rx_bytes_per_second>%" PRIu64 "</rx_bytes_per_second>", (uint64_t)(pdata.rxpercentile / (double)300));
-	printf("<tx_bytes_per_second>%" PRIu64 "</tx_bytes_per_second>", (uint64_t)(pdata.txpercentile / (double)300));
-	printf("<total_bytes_per_second>%" PRIu64 "</total_bytes_per_second>", (uint64_t)(pdata.sumpercentile / (double)300));
+	printf("<rx_bytes_per_second>%" PRIu64 "</rx_bytes_per_second>", (uint64_t)((double)pdata.rxpercentile / (double)300));
+	printf("<tx_bytes_per_second>%" PRIu64 "</tx_bytes_per_second>", (uint64_t)((double)pdata.txpercentile / (double)300));
+	printf("<total_bytes_per_second>%" PRIu64 "</total_bytes_per_second>", (uint64_t)((double)pdata.sumpercentile / (double)300));
 	printf("</95th_percentile>\n");
 
 	printf("  </bandwidth>\n");
